Clean up code in the Settings Activity
Extract all strings to the resource files (helpful for translations in 
the future), Move and simplify functions.
