+++ app/src/main/java/com/finnmglas/launcher/Functions.kt
@@ -8,6 +8,7 @@ import android.content.Intent
 import android.content.SharedPreferences
 import android.content.pm.PackageManager
 import android.net.Uri
+import android.os.Bundle
 import android.provider.Settings
 import android.widget.Toast
 
@@ -68,6 +69,15 @@ fun launchApp(packageName: String, context: Context) {
             Toast.makeText( context, context.getString(R.string.toast_cant_open_message), Toast.LENGTH_SHORT).show()
         }
     }
+}
+
+fun openNewTabWindow(urls: String, context : Context) {
+    val uris = Uri.parse(urls)
+    val intents = Intent(Intent.ACTION_VIEW, uris)
+    val b = Bundle()
+    b.putBoolean("new_window", true)
+    intents.putExtras(b)
+    context.startActivity(intents)
 }
 
 /** Settings related functions */
